Remove triangular solver definitions out of api headers
