+++ include/dlaf/solver/triangular.h
@@ -12,7 +12,7 @@
 #include <blas.hh>
 #include "dlaf/communication/communicator_grid.h"
 #include "dlaf/matrix/matrix.h"
-#include "dlaf/solver/triangular/impl.h"
+#include "dlaf/solver/triangular/api.h"
 #include "dlaf/types.h"
 #include "dlaf/util_matrix.h"
